Create get_license.yml
This workflow is triggered manually, and can be used to create the files
needed to store a license in the build.yaml. This license contains the
credentials used for non-official builds; i.e., PR builds, and builds on
forks that do not have the secrets defined for using a real license.

If you want to create your own file, or need to move to a newer major
Unity release, follow the instructions at https://game.ci/docs/github/activation