Sketch metadata now supports fallback brush GUIDs.
This updates the metadata schema version to 3.

The schema now has a FallbackBrushIndex guid array that stores the guid
of the built-in brush that should be used if the requested brush cannot
be found.

In the case the brush can't be found, a new brush descriptor is created
by duplicating the descriptor of the fallback brush and changing its
guid to match the missing brush. In this way the scene can be saved out
again without the strokes of the missing brush being affected.