Sign all builds that are built from pushes to main (#64)
This requires the addition of 4 secrets:

ANDROID_KEYSTORE_BASE64 -- the output of `base64 -i foo.keystore`
ANDROID_KEYSTORE_PASS -- the password for the keystore
ANDROID_KEYALIAS_NAME -- the name of the alias
ANDROID_KEYALIAS_PASS -- should match KEYSTORE_PASS in our case

Because it uses Github secrets, it won't run on a fork, so we won't do
this for PRs. But anything coming from the main tree can use this.

We do still keep the github name in the package name, to differentiate
between a formal release and an unreleased candidate.