Create release upon each build and standardize version numbers
Each push to a PR will create builds (as artifacts only) with the
version number of X.Y.Z and a stamp of gABCDEFA, where X.Y is the last
tag in the main repo of the form vX.Y, Z is the number of commits since
then, and ABCDEFA is the git hash of the tip of PR

Each push to main (icosa-gallery only) will create artifacts of the form
X.Y.Z (same meaning as above), as well as an easier to use 'pre-release'
release on the github releases page. It will also tag the commit with
X.Y.Z (no 'v') to make it easier to cross-reference.

Each push of a tag of the format vX.Y to main will create an official release named X.Y.0.

Both of these releases include the same 4 files; Windows + Oculus, regular +
Experimental. The other artifacts (Mac OS X, Linux, Windows Monoscopic)
remain available from the Artifacts page, but are not exposed on the
'releases' link.

All Oculus builds are signed, and the android-specific bundleVersionCode
is now updated automatically (1000000*X+1000*Y+Z).

Creating a new formal release requires two steps:
* Update the following three files, create a PR as usual, and merge it:
  Assets/Scenes/Main.unity:  m_VersionNumber: X.Y.0
  ProjectSettings/ProjectSettings.asset:  bundleVersion: X.Y.0
  README.md:## Current Version: `X.Y.0`
* Tag the merge commit with `vX.Y` (do not include the .0), and push the
  tag to trigger the actual build of X.Y.0.