Use secrets.UNITY_EMAIL/_PASSWORD/_SERIAL if present, otherwise use hardcoded free license
If Github Secrets UINTY_SERIAL is defined, then the build actions will
be run with the value of Github Secrets UNITY_SERIAL, UNITY_EMAIL, and
UNITY_PASSWORD.

If not, it will use the free license included in this file.

This means that builds from the main repo itself can use a protected
professional license, whereas PRs from forks (which do not have access
to the github secrets, for security reasons (to avoid leaking
credentials)) can still be built, albeit with a free license (i.e., with
a splash screen).