Name builds based on the source (#117)
In order to make it easier to run multiple versions, instead of a fixed
"-github" name (and package name of com.icosa.openbrushgithub), this
commit changes the logic for CI builds.

* Builds from PRs are named PR#xx (PRxx in the package name)
* Builds from branches (triggered with a [CI BUILD] tag) are named based on the branch
name. Special characters of /[#_-]/ are removed from the package name
* All commits on main have the same name, whether they're pre-release
builds or formal release builds. This means that the pre-release builds
will NOT have a unique name, but this will allow us to upload them to a
"beta" release channel