Manually install dotnet-format, pinned to the previous version (#166)
The latest dotnet-format is 7.0.x and does not run with .NET 3.1.2 or 5.x, which is the stable release. We could move to a preview release, but there are two open issues on dotnet format; see 1317 and 1318 at their repo.

However, the way pre-commit handles dotnet packages, it always installs the latest version (dotnet tool install ...). There is no way to specify a specific version.

So we have this charming hack to trick pre-commit into using our manually installed version.