Disable custom keystore in repo; enable it for all CI builds
In #217, it was enabled in the ProjectSettings file. However, this broke
local compilation, since the keystore was not available for repo
clones/forks.

It's better to leave it disabled, and just enabled it in all CI builds
(where we use either the real keystore for formal builds, or our fake
one for PRs)