Continue uploading builds to Steam's '*-experimental' branches
After PR#344, the same build serves both regular mode and experimental
mode, but people who are on either beta-experimental (the new name) or
prerelease-experimental (the old name, see PR#291 and #315 for details)
should get this build too.

Once 2.0 is released, we'll delete all the channels except "beta" and
everyone will default to the formal release.