Support hotfixes / maintenance releases properly
Currently, there is an unwritten assumption in the code that tags of the
form vX.Y will always be on main.

This PR allows those tags to be on a branch, and it will correctly
create the store builds, the Github release, and calculate version codes
for subsequent builds. The naming convention will continue to be X.Y.0
for formal builds, whether from main or from a branch, and X.Y.Z for
pre-release builds, where Z represents the number of commits (merge
commits or squashed commits, not commits in a feature branch) that are
in this tag but were not in X.Y.0. As such, if a maintenance release is
created, the next pre-release version will have Z increase by 2; one for
the latest feature, and one for the merge of the maintenance branch.

One requirement is that this branch must be merged to main using a merge
commit, not a squash, in order for the version numbers to be calculated
correctly. There should be zero conflicts when this merge is done, but
if there are any, the version in main should take priority, assuming
that the fixes in the branch were also commited there. If this is not
true, then the conflict must be resolved manually.