Correctly calculate changelogs when a hotfix has been made
When a hotfix was created, it means that the last vX.Y tag was not a
--first-parent of the current tip. Since the hotfix (probably) just had
cherry-picks, we want to include those PRs (as well as those that were
not cherry-picked when releasing the next, formal, version. To do this,
we need to go back to the last vX.Y tag that was made on a commit on
main.