Webcam (or other streaming video source) panel (#477)
Suggested by Shawn Chiki to help with performances where a view of the outside world is essential. I suspect it might be useful for streamers as well. Use with OBS allows practically anything to be streamed into the headset.

(Not supported on any Quest headsets due to lack of support for webcam devices. Not sure about other standalone headsets)