Enable masking of UNITY_PASSWORD (#566)
This is NOT a security feature! The value stored in the file in the clear is still stored there! This merely hides the print in the builder action, making it slightly less obvious. (this only applies to runs from forks; from the repo, it's always masked because it comes from a secret)