Retry build on license acquisition failure (#571)
Reverts #570

The unity-builder action now exposes the exit code of the call to Unity.
License failure results in an exit code of 1. In this case, call a new
step to try to rerun it.

It seems like no other errors return 1. I tried both a syntax error and
an invalid package. Both returned 133.