Change build directories to align with the unity-builder standard directories (#593)
We'll build to ~/build/$vrsdk/$platform; we tell unity-builder that we're starting in ~/build/$vrsdk. The artifacts now just have $platform, instead of $platform-$vrsdk as the leading directory. Stores are unaffected.

This lets us use the chmod from the unity-builder action, as well as the built in ChownFilesTo instead of our own invocation of a docker container.