Fix broken Dependabot PRs (#650)
Normally, Github has a security model whereby secrets and vars are
accessible to PRs from a branch in the main repo, but not to PRs from
forks. This is to prevent a malicious PR from exfiltrating secrets. Why
this should apply to vars (which are not masked) is unclear, but that's
how it works.

Dependabot PRs, even though they're from the main repo, do not have
access to secrets, for the same reason. See
https://github.com/dependabot/dependabot-core/issues/3253#issuecomment-852541544
. However, for some reason, dependabot PRs *do* have access to vars!

To prevent this from breaking Android signing, we'll move
ANDROID_KEYALIAS_NAME back to a secret, instead of a variable, even
though it's not intrinsically secret.