Always specify --age-group for Oculus (#670)
Without this, Beta builds show a warning in the dashboard, and LIVE
builds do not get published.

The doc at https://developer.oculus.com/resources/age-groups/ suggests
that this is per-app, but the docs at
https://developer.oculus.com/resources/publish-reference-platform-command-line-utility/
refer to it as a per-build setting. It seems it's actually per build.

It also has to be set for Q1 only builds, despite the fact that Q1
doesn't officially support this feature!