Workaround for PCVR depth submission (#685)
This is a partial fix for the issue discussed in #420

The issue was coming from RenderWrapper.cs, where an offscreen render texture is created so it can be reused and rescaled for performance when using the video recording tools. I'm not entirely sure why this is breaking our XR pipeline, as it appears the render texture is created properly.

As a stopgap, we keep using the normal camera, but switch to the offscreen render during recording.

This may be magically resolved when we upgrade Unity versions.