Add support for a [xCI BUILD DEV] (without the x) to build Development builds
Note that this does not need to also contain the regular [xCI BUILDx]
command, and indeed, if this is part of a pull request, it should not
(as then it will build twice).

Also note that we currently have 11 builds, but only 20 available
workers, which means that enabling Development builds will require a
second round of builds. The queuing system will handle this
automatically, but it will take longer.

This commit also adds a dummy list of names to prevent the includes from
overwriting each other. This is a quirk of Github matrices, whereby if
there's *just* an include: list, it will copy them as-is, but once
there's a predefined matrix, then include will overwrite, rather than
append, anything with a matching key. By priming the list with the
names, even without any of the config, we ensure that none will be
skipped, and we just use the include list for the configurations.