Lock panels to beginner mode when a user enters a multiplayer room
To block the ability to join a multiplayer session when in advance mode: I've changed the availability of the Globalcommand.MultiplayerJoinRoom to only be available when the user is beginner mode.  As part of this I also changed the MultiplayerPanel.Prefab button (join room) to AllowUnavailable to True. This change is also supported by a check that happens every time the multiplayer panel is open and a UI alert message explaining to switch to beginner mode.

To block the ability to switch to advance mode during a multiplayer session: I've changed the availability of the Globalcommand.AdvancedPanelsToggle in the IsCommandAvailable() essentially I check if the user is in a multiplayer room and if it is the advancePanelsToggle becomes unvalaible. As part of this I also changed the Admin panel prefab buttons (advance beginner mode) to AllowUnavailable to True.