Fix the mirror for multiplayer
Fix the mirror widget  for multiplayer

The first problem is related to the fact that the current implementation of the multiplayer does not properly transmit and serialize the chain of commands especially the property of the basecommand m_Children and m_Parent.

The second problem this commit address is the fact that In mirror mode, the brush stroke command is invoked twice. Initially, a single brush stroke command is sent. The second invocation sends a command chain, including the initial brush stroke as a parent and a second mirrored stroke as its child. Sending both instances through PhotonRPC causes issues, as the first brush stroke command is received twice, leading to duplication and processing errors on the receiving peer. To resolve this  we use PerformAndRecordCommand in place of RecordCommand, and we change the PointerManager.FinalizeLine(), PointerScript.DetachLine(), SketchMemoryScript.MemorizeBatchedBrushStroke() to utilize a parameter