Improving Remote Player Registration
Remote players were not correctly registered when already present in the room because `INetworkRunnerCallbacks.OnPlayerJoined` in Fusion 1 is not called for users who are already in the room.

To address this, we added and registered existing users using the `CheckExistingUsers()` method.

Additionally, in `MultiplayerManager.cs`, the local player ID was not registered correctly, which has now been fixed.

A display for both the local and remote user IDs was added to `MultiplayerManagerEditor.cs` to facilitate debugging.