Implement Estimation of Command-to-Network-Message Conversion
Some commands must be split into multiple network messages due to payload size limitations per message. In this changeset, we implemented a function to estimate the number of network messages required to transmit each command. This enhancement improves the efficiency of the coroutine responsible for synchronizing the scene with new users.