Fix to RemoveKeyFrame and Changed AddLayerRefresh
Fixed the RemoveKeyFrame by adding a ClearLayerContent in it. Without it, SketchWriter.cs:EnumerateAdjustedSnapshots() will fail to reconcile active canvases and all strokes.

Changed AddLayerRefresh so that the added frame from SceneScript is set to the front of the timeline. Also will return the canvas where the slider is on, so when you draw, you're drawing on the right frame.