Prepare Remote Player List for Display
Introduce a `RemotePlayer` class with `PlayerId`,  `Nickname` and `ItransientData`, extendable with Role (e.g., observer/editor/owner) and VoiceStatus (e.g., muted/active speaker). Abstracts player details from the Photon SDK for better flexibility and clarity.
Includes a list of `RemotePlayer` ready to be displayed and interacted with on a panel.